une asser propre qui marche

+++ pid_f4/Inc/odometry.h
@@ -10,12 +10,12 @@
 typedef struct Odometry_S{
   Encoder encoder_l;
   Encoder encoder_r;
-  volatile float x;
-  volatile float y;
-  volatile float theta;
-  float speed_l;
-  float speed_r;
-  float speed_percent_tolerance;
+  volatile double x;
+  volatile double y;
+  volatile double theta;
+  double speed_l;
+  double speed_r;
+  double speed_percent_tolerance;
 }Odometry;
 
 void init_odometry(Odometry *odometry, TIM_HandleTypeDef *htim_l, TIM_HandleTypeDef *htim_r, TIM_HandleTypeDef *htim_poll);
